test/web/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.4.3
psycopg2-binary==2.8.4
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
PyJWT==1.7.1